+++ src/main/java/com/haibao/business/domain/enums/SubjectType2Code.java
@@ -5,19 +5,19 @@ package com.haibao.business.domain.enums;
  * Created by haibao on 2018/1/22.
  */
 public enum SubjectType2Code {
-    PUBLIC((byte) 0, "公共课"),
-    SPECIALIZED((byte) 1, "专业课"),
-    BASE_SPECIALIZED((byte) 2, "专业基础课");
+    PUBLIC( 0, "公共课"),
+    SPECIALIZED( 1, "专业课"),
+    BASE_SPECIALIZED( 2, "专业基础课");
 
-    SubjectType2Code(byte code, String desc) {
+    SubjectType2Code(int code, String desc) {
         this.code = code;
         this.desc = desc;
     }
 
-    private byte code;
+    private int code;
     private String desc;
 
-    public byte code() {
+    public int code() {
         return code;
     }
